feat: Update sign-in and sign-up forms
This commit updates the sign-in and sign-up forms to include the "name" field. Previously, only the "email" and "password" fields were present. The addition of the "name" field allows users to provide their full name during the sign-up process. This information will be used for personalization purposes within the application.

Code changes:
- Added "name" field to the sign-in and sign-up forms
- Updated the form submission logic to include the "name" field in the request payload